requirements.txt
Flask
authlib
gunicorn
requests
google-cloud-firestore
google-auth
google-auth-oauthlib
google-cloud-storage>=2.0.0
python-dateutil>=2.8.0
pillow
google-cloud-tasks
google-cloud-secret-manager
